pyproject.toml
[project]
name = "books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "pillow>=11.1.0",
    "reflex>=0.7.1",
    "wikipedia-api>=0.8.1",
]
